Allow record priority to be specified when creating and updating
Some record types (like MX) allow a priority to be specified, and it's
useful to have the option to provide it.

Like ttl, the prio param is optional so I've emulated that way ttl is
parsed and handled.

I've also added Ttl and Prio to the test cases to ensure conditional
code is executed.